Changelog,
-Major changes to all code, most all objects are now controlled in vectors for easier code management, main pipelines for game have been layed time to build out the content

+++ SFML Console Chat/Player.h
@@ -10,8 +10,8 @@ public:
 
 	//Leveling
 	int level;
-	float exp;
-	float expNext;
+	int exp;
+	int expNext;
 	int sp;
 
 	//Attributes
